+++ busker-amp.py
@@ -7,6 +7,7 @@ import threading
 from modules.bt_footswitch_recv import ESP32BLEClient
 import logging
 import asyncio
+from display_utils import display_loop
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 s = pyo64.Server()
@@ -221,82 +222,19 @@ async def run_pedal_loop():
        await asyncio.sleep(0.07)  # Keep the event loop running
 
 
-def clear_screen():
-    """Clear the console screen"""
-    os.system('cls' if os.name == 'nt' else 'clear')
-
-def move_cursor_to_top():
-    """Move cursor to top of console without clearing"""
-    if os.name == 'nt':  # Windows
-        os.system('cls')
-    else:  # Unix/Linux/Mac
-        print('\033[H', end='')
-
-def display_values():
-    """Display the current values in a table format"""
-    with data_lock:
-        current_fx = fx.copy()
-        current_looper = looperState
-        current_old_looper = oldLooperState
-    
-    # Move cursor to top
-    move_cursor_to_top()
-    
-    # Create the display
-    print("=" * 60)
-    print("           AUDIO EFFECTS MONITOR")
-    print("=" * 60)
-    print()
-    
-    # Main effects table
-    print("┌─────────────┬─────────┬─────────────┬─────────┐")
-    print("│ PARAMETER   │  VALUE  │ PARAMETER   │  VALUE  │")
-    print("├─────────────┼─────────┼─────────────┼─────────┤")
-    print(f"│ WET         │ {current_fx['wet']:7.3f} │ DELAY       │ {current_fx['delay']:7.3f} │")
-    print(f"│ DRY         │ {current_fx['dry']:7.3f} │ REVERB      │ {current_fx['reverb']:7.3f} │")
-    print(f"│ BASS        │ {current_fx['bass']:7.3f} │ DISTORT     │ {current_fx['distort']:7.3f} │")
-    print(f"│ MID         │ {current_fx['mid']:7.3f} │ WAH         │ {current_fx['wah']:7.3f} │")
-    print(f"│ TREBLE      │ {current_fx['treb']:7.3f} │             │         │")
-    print("└─────────────┴─────────┴─────────────┴─────────┘")
-    print()
-    
-    # Looper state
-    print("┌─────────────────────────────────────────────────┐")
-    print("│                 LOOPER STATUS                   │")
-    print("├─────────────────────────────────────────────────┤")
-    print(f"│ Current State: {current_looper:15s}         │")
-    print(f"│ Previous State: {current_old_looper:14s}         │")
-    print("└─────────────────────────────────────────────────┘")
-    print()
-    
-    # Add some padding to prevent screen flickering
-    print("\n" * 5)
-    print("Press Ctrl+C to exit...")
-
-def display_loop():
-    """Continuously update the display"""
-    # Clear screen once at start
-    clear_screen()
-    
-    while True:
-        try:
-            display_values()
-            time.sleep(0.1)  # Update display 10 times per second
-        except KeyboardInterrupt:
-            break       
-
-
-
 if __name__ == "__main__":
-    # Create two threads
+    # Create threads
     thread1 = threading.Thread(target=inputLoop)
     thread2 = threading.Thread(target=controlLoop)
-    thread3 = threading.Thread(target=display_loop, daemon=True)
+    thread3 = threading.Thread(
+        target=display_loop, 
+        args=(fx, looperState, oldLooperState, data_lock), 
+        daemon=True
+    )
 
-    # Start the threads
     thread1.start()
     thread2.start()
-    thread3.start()  # Start the display thread
+    thread3.start()
     try:
         asyncio.run(run_pedal_loop())
         # Join the threads to the main thread to keep them running
